modified:   .firebase/hosting.YnVpbGRcd2Vi.cache 	modified:   lib/data_layer/mempool.dart 	modified:   lib/data_layer/robot_database.dart 	modified:   lib/data_layer/tx_database.dart 	modified:   lib/domain_layer/account.dart 	modified:   lib/domain_layer/blockchain.dart 	modified:   lib/domain_layer/transaction.dart 	modified:   lib/presentation_layer/screens/confirmed_operations_screen.dart 	modified:   lib/presentation_layer/screens/pending_operations_screen.dart 	modified:   lib/presentation_layer/screens/transfer_rights_screen.dart 	modified:   lib/presentation_layer/widgets/transactions_table.dart 	modified:   server_example/bin/blockchain/account.dart 	modified:   server_example/bin/blockchain/blockchain.dart 	modified:   server_example/bin/blockchain/signature.dart 	modified:   server_example/bin/blockchain/transaction.dart 	modified:   server_example/bin/database/mempool.dart 	modified:   server_example/bin/database/robot_database.dart 	modified:   server_example/bin/database/tx_database.dart 	modified:   server_example/bin/server.dart

+++ lib/domain_layer/account.dart
@@ -65,8 +65,9 @@ class Account {
         await blockchain.robotDatabase.accountExists(accountID);
     if (!accountIsValid) return false; // returning null if not
 
-    // getting robots from robotDatabase
-    Set<Robot> robots = await blockchain.robotDatabase.getRobots(accountID);
+    // getting robots from robotDatabase. robots from mempool are not included
+    Set<Robot> robots = await blockchain.robotDatabase
+        .getRobots(accountID, includeMempool: false);
 
     verifiedAccount = Account(
       accountID: accountID,
@@ -75,6 +76,12 @@ class Account {
     );
 
     return true;
+  }
+
+  /// Returns `true` if the account with given ID already exists in `robotDatabase`.
+  static Future<bool> exists(String accountID) async {
+    bool exists = await blockchain.robotDatabase.accountExists(accountID);
+    return exists;
   }
 
   /// Adds the given robot to the account.
@@ -106,7 +113,7 @@ class Account {
     }
     try {
       // removing a robot from robotDatabase
-      await blockchain.robotDatabase.removeRobot(robot);
+      if (updateDB) await blockchain.robotDatabase.removeRobot(robot);
       // updating current state
       robots.remove(robot);
     } catch (e) {
@@ -169,6 +176,16 @@ class Account {
   /// Returns all confirmed operations, which were performed by the account.
   Future<List<Operation>> getConfirmedOperations() async {
     return await blockchain.txDatabase.getAccountOperations(accountID);
+  }
+
+  /// Returns stream of pending operations, which were performed by the account.
+  Stream<List<Operation>> getPendingOperationsStream() {
+    return blockchain.mempool.getAccountOperationsStream(accountID);
+  }
+
+  /// Returns stream of sconfirmed operations, which were performed by the account.
+  Stream<List<Operation>> getConfirmedOperationsStream() {
+    return blockchain.txDatabase.getAccountOperationsStream(accountID);
   }
 
   // equivalent of printBalance()
